Extract gson file
