+++ src/main/kotlin/ffc/entity/gson/Exclude.kt
@@ -1,0 +1,19 @@
+package ffc.entity.gson
+
+import com.google.gson.ExclusionStrategy
+import com.google.gson.FieldAttributes
+
+@Retention(AnnotationRetention.RUNTIME)
+@Target(AnnotationTarget.FIELD)
+annotation class Exclude
+
+class ExcludeAnnotationStrategy : ExclusionStrategy {
+
+    override fun shouldSkipField(f: FieldAttributes): Boolean {
+        return f.getAnnotation(Exclude::class.java) != null
+    }
+
+    override fun shouldSkipClass(clazz: Class<*>): Boolean {
+        return false
+    }
+}
