+++ src/main/kotlin/ffc/entity/gson/Gson.kt
@@ -20,21 +20,10 @@ package ffc.entity.gson
 import com.fatboyindustrial.gsonjodatime.DateTimeConverter
 import com.fatboyindustrial.gsonjodatime.LocalDateConverter
 import com.fatboyindustrial.gsonjodatime.LocalDateTimeConverter
-import com.google.gson.ExclusionStrategy
-import com.google.gson.FieldAttributes
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
-import com.google.gson.JsonDeserializationContext
-import com.google.gson.JsonDeserializer
-import com.google.gson.JsonElement
-import com.google.gson.JsonSerializationContext
-import com.google.gson.JsonSerializer
 import com.google.gson.reflect.TypeToken
 import ffc.entity.Identity
-import ffc.entity.THAI_CITIZEN_ID
-import ffc.entity.THAI_HOUSEHOLD_ID
-import ffc.entity.ThaiCitizenId
-import ffc.entity.ThaiHouseholdId
 import ffc.entity.User
 import me.piruin.geok.LatLng
 import me.piruin.geok.geometry.Geometry
@@ -54,7 +43,7 @@ val ffcGson: Gson by lazy {
             .create()
 }
 
-private fun GsonBuilder.adapterForExtLibrary(): GsonBuilder {
+internal fun GsonBuilder.adapterForExtLibrary(): GsonBuilder {
     adapterFor<Geometry>(GeometrySerializer())
     adapterFor<LatLng>(LatLngSerializer())
     adapterFor<DateTime>(DateTimeConverter())
@@ -73,46 +62,3 @@ fun Any.toJson(gson: Gson = ffcGson): String = gson.toJson(this)
 
 inline fun <reified T> String.parseTo(gson: Gson = ffcGson): T = gson.fromJson(this, typeTokenOf<T>())
 
-class IdentityJsonAdapter : JsonDeserializer<Identity>, JsonSerializer<Identity> {
-    override fun serialize(src: Identity, typeOfSrc: Type?, context: JsonSerializationContext): JsonElement {
-        return context.serialize(src)
-    }
-
-    override fun deserialize(json: JsonElement, typeOfT: Type, context: JsonDeserializationContext): Identity {
-        val jsonObj = json.asJsonObject
-        return when (jsonObj.get("type").asString) {
-            THAI_CITIZEN_ID -> ThaiCitizenId(jsonObj.get("id").asString)
-            THAI_HOUSEHOLD_ID -> ThaiHouseholdId(jsonObj.get("id").asString)
-            else -> throw IllegalArgumentException("Not support Identity type")
-        }
-    }
-}
-
-class UserJsonAdapter() : JsonSerializer<User> {
-
-    val gson = GsonBuilder()
-            .setExclusionStrategies(ExcludeAnnotationStrategy())
-            .adapterForExtLibrary()
-            .create()
-
-    override fun serialize(user: User, typeOfSrc: Type, context: JsonSerializationContext): JsonElement {
-        return gson.toJsonTree(user).apply {
-            if (!user.isTempId) asJsonObject.remove("password")
-        }
-    }
-}
-
-@Retention(AnnotationRetention.RUNTIME)
-@Target(AnnotationTarget.FIELD)
-annotation class Exclude
-
-class ExcludeAnnotationStrategy : ExclusionStrategy {
-
-    override fun shouldSkipField(f: FieldAttributes): Boolean {
-        return f.getAnnotation(Exclude::class.java) != null
-    }
-
-    override fun shouldSkipClass(clazz: Class<*>): Boolean {
-        return false
-    }
-}
